Move BreakLongEdges into phase3 package as top-level func
